+++ src/neuroconv/baseextractorinterface.py
@@ -1,4 +1,5 @@
 """Abstract class defining the structure of all Extractor-based Interfaces."""
+
 from abc import ABC
 from typing import Optional
 
@@ -16,6 +17,10 @@ class BaseExtractorInterface(BaseTemporalAlignmentInterface, ABC):
     ExtractorModuleName: Optional[str] = None
     ExtractorName: Optional[str] = None  # Defaults to __name__.replace("Interface", "Extractor").
     Extractor = None  # Class loads dynamically on first call to .get_extractor()
+    help: Optional[str] = None  # Help string for the extractor. If defined, this is used for the help message of GUIDE.
+    display_name: Optional[str] = (
+        None  # Display name for the extractor. If defined, this is used for the display name of GUIDE.
+    )
 
     @classmethod
     def get_extractor(cls):