+++ tests/test_minimal/test_tools/test_backend_and_dataset_configuration/test_helpers/test_configure_backend_defaults.py
@@ -1,12 +1,14 @@
 """Unit tests for `get_default_dataset_configurations`."""
+
 from pathlib import Path
-from typing import Literal
+from typing import Literal, Tuple
 
 import numcodecs
 import numpy as np
 import pytest
 from hdmf.common import DynamicTable, VectorData
 from hdmf.data_utils import DataChunkIterator
+from numpy.testing import assert_array_equal
 from pynwb.testing.mock.base import mock_TimeSeries
 from pynwb.testing.mock.file import mock_NWBFile
 
@@ -16,6 +18,24 @@ from neuroconv.tools.nwb_helpers import (
     configure_backend,
     get_default_backend_configuration,
 )
+
+
+@pytest.fixture(scope="session")
+def integer_array(
+    seed: int = 0,
+    dtype: np.dtype = np.dtype("int16"),
+    shape: Tuple[int, int] = (30_000 * 5, 384),
+):
+    """
+    Generate an array of integers.
+
+    Default values are chosen to be similar to 5 seconds of v1 NeuroPixel data.
+    """
+    random_number_generator = np.random.default_rng(seed=seed)
+
+    low = np.iinfo(dtype).min
+    high = np.iinfo(dtype).max
+    return random_number_generator.integers(low=low, high=high, size=shape, dtype=dtype)
 
 
 @pytest.mark.parametrize(
@@ -29,10 +49,14 @@ from neuroconv.tools.nwb_helpers import (
 )
 @pytest.mark.parametrize("backend", ["hdf5", "zarr"])
 def test_simple_time_series(
-    tmpdir: Path, case_name: str, iterator: callable, iterator_options: dict, backend: Literal["hdf5", "zarr"]
+    tmpdir: Path,
+    integer_array: np.ndarray,
+    case_name: str,
+    iterator: callable,
+    iterator_options: dict,
+    backend: Literal["hdf5", "zarr"],
 ):
-    array = np.zeros(shape=(30_000 * 5, 384), dtype="int16")
-    data = iterator(array, **iterator_options)
+    data = iterator(integer_array, **iterator_options)
 
     nwbfile = mock_NWBFile()
     time_series = mock_TimeSeries(name="TestTimeSeries", data=data)
@@ -57,14 +81,16 @@ def test_simple_time_series(
         elif backend == "zarr":
             assert written_data.compressor == numcodecs.GZip(level=1)
 
+        assert_array_equal(x=integer_array, y=written_data[:])
+
 
 @pytest.mark.parametrize("backend", ["hdf5", "zarr"])
-def test_simple_dynamic_table(tmpdir: Path, backend: Literal["hdf5", "zarr"]):
-    data = np.zeros(shape=(30_000 * 5, 384), dtype="int16")
-
+def test_simple_dynamic_table(tmpdir: Path, integer_array: np.ndarray, backend: Literal["hdf5", "zarr"]):
     nwbfile = mock_NWBFile()
     dynamic_table = DynamicTable(
-        name="TestDynamicTable", description="", columns=[VectorData(name="TestColumn", description="", data=data)]
+        name="TestDynamicTable",
+        description="",
+        columns=[VectorData(name="TestColumn", description="", data=integer_array)],
     )
     nwbfile.add_acquisition(dynamic_table)
 
@@ -87,3 +113,5 @@ def test_simple_dynamic_table(tmpdir: Path, backend: Literal["hdf5", "zarr"]):
             assert written_data.compression == "gzip"
         elif backend == "zarr":
             assert written_data.compressor == numcodecs.GZip(level=1)
+
+        assert_array_equal(x=integer_array, y=written_data[:])